Implement Vertical and Lateral Movement
Implement basic vertical(j) and lateral(q/e) movement for the person object. This included defining new variables - isJumping, moveLeft, moveRight - in the constructor and adding event listeners for key presses.

The important change here that might impact other developers is my change to draw_person call in display(). We initially hardcoded the coordinates(last parameter), but this was resetting the lateral movement each render. To prevent this, I defined a new variable - person_transform - in the constructor that allows changes in coordinates to persist in state. This also allows for easier movement of the person object, no longer having to loop through each appendage.