Perspective change -> long walkway
Changed the perspective so that the walkway is rendered as one long rectangle that the
player runs across. Some of the code that was meant for the other perspective can still be used
so I kept it in but commented it out. Changed some of the initialization matrices of objects
so they are flush to the new walkway

Add some state variables for collision detection - not done yet.