Flyer person now turns
Define a map that keeps track of each flyerperson's relative movement in the x direction and if that person has turned or not. These allow the flyerperson to detect if the player is within a certain y-range. If so, the flyer person will walk into the path and turn to face the player.

To-do: Make it so the turn is gradual