Update set_enemies function
update set_enemies function so that each call will remove all key/value pairs from the starship_locations map. This means that when you move on to a new level, the old starships won't be stored anywhere and won't be displayed on the map. Will soon do something similar with flyerperson2