Add flyerpersons to map
Add flyer persons to map but only after they have started moving which is when their coordinates are reflected in the flyerperson_info map. Not sure if I will keep it this way as it isn't very useful to the player. If not, will need other state variable to hold coordinates.